apiDir/requirements.txt
asgiref==3.2.10
astroid==2.4.1
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.1
django-extensions==3.0.5
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.11.1
djangorestframework-datatables==0.5.2
gunicorn==20.0.4
idna==2.10
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
psycopg2==2.8.5
Pygments==2.6.1
pylint==2.5.2
pytz==2020.1
requests==2.24.0
simplejson==3.17.2
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
whitenoise==5.2.0
wrapt==1.12.1
